Add price, SMA and chart nodes with realtime update
